API

API

Limbdark.transit_init — Method.

transit_init(r,b,u_n,grad)

Initialize and return a transit structure. This structure holds user-facing information about the transit as well as internal temporary variables used for computing the light curve. Pass an instance of this structure to transit_poly() to compute the actual flux.

Arguments

r::Real: The radius of the occultor in units of the radius of the occulted body.

b::Real: The (initial) impact parameter of the occultation.

u_n::Array{Real,1}: The array of limb darkening coefficients.

grad::Bool: Compute the gradient of the flux as well?

source

Limbdark.transit_poly — Method.

transit_poly(r,b,u_n)

Given a radius-ratio, impact parameter, vector of limb-darkening coefficients of size N, returns the flux (normalized to one for unocculted star) for a limb-darkened transit.

Arguments

r::Real: The radius of the occultor in units of the radius of the occulted body.

b::Real: The (initial) impact parameter of the occultation.

u_n::Array{Real,1}: The array of limb darkening coefficients. Size N.

source

Limbdark.compute_uniform! — Method.

compute_uniform(t)

Computes the coefficient for the uniform disk case, s_0, given a TransitStruct instance t.

source

Limbdark.s2! — Method.

s2!(r,b,s2_grad)

Computes the linear limb-darkening case, as well as the gradient, s2_grad=[ds_2/dr,ds_2/db] is a pre-allocated two-element array. Returns s2 and complete elliptic integrals, E(k) and (E(m)-(1-m)K(m))/m, where m = k^2.

Example

julia> s2_grad=[0.0,0.0]
julia> s2!(0.1,0.5,s2_grad)
(2.067294367278038, 1.4726447391114554, 0.8111640472029077)
julia> s2_grad
2-element Array{Float64,1}:
 -0.53988  
  0.0182916

source

Limbdark.s2 — Method.

s2(r,b)

Compute the integral of the linear limb-darkening case.

Examples

julia> s2(0.1,0.5)
2.067294367278038

source

Limbdark.s2_ell — Method.

s2_ell(r,b)

Compute the integral of the linear limb-darkening case, s_2 from starry (which is S_1 in Agol & Luger) and also return two elliptic integrals, E(k) and (E(m)-(1-m)K(m))/m, where m = k^2.

Examples

julia> s2_ell(0.1,0.5)
(2.067294367278038, 1.4726447391114554, 0.8111640472029077)

source

Limbdark.sqarea_triangle — Method.

sqarea_triangle(a,b,c)

Function which computes sixteen times the square of the area of a triangle with sides a, b and c using Kahan method. How to compute (sixteen times the) area squared of triangle with high precision (Goldberg 1991).

source

Limbdark.transit_poly! — Method.

transit_poly!(t)

Given a TransitStruct instance t, with Green's coefficients already initialized, computes a limb-darkened transit light curve with optional gradient with respect to r, b and u coefficients.

source

Limbdark.transit_poly! — Method.

transit_poly!(r,b,u_n,dfdrbu)

Given a radius-ratio, impact parameter, vector of limb-darkening coefficients of size N, and pre-allocated vector for derivatives of size N+2, returns the flux (normalized to one for unocculted star) for a limb-darkened transit light curve and its gradient with respect to the input parameters is returned.

Arguments

r::Real: The radius of the occultor in units of the radius of the occulted body.

b::Real: The (initial) impact parameter of the occultation.

u_n::Array{Real,1}: The array of limb darkening coefficients. Size N.

dfdrbu::Array{Real,1}: Gradient of the flux with respect to all input parameters. Size is N+2

source

Limbdark.transit_poly_g! — Method.

transit_poly_g!(t)

Given a TransitStruct instance t, with Green's coefficients already initialized, computes a limb-darkened transit light curve including gradient with respect to r, b and g coefficients.

source

Limbdark.transit_poly_g — Method.

transit_poly_g(t)

Given a TransitStruct instance t, computes a limb-darkened transit light curve without gradient.

source

Limbdark.Transit_Struct — Type.

Transit_Struct

Structure to hold arrays and other quantities for computing transit.

Members

r::T: radius ratio

b::T: impact parameter

u_n::Array{T,1}: limb-darkening coefficients

n::Int64: number of limb-darkening coefficients

n_max::Int64: maximum number of terms in M_n

g_n::Array{T,1}: Green's basis coefficients

sn::Array{T,1}: Green's basis terms

Mn::Array{T,1}: integral over $(k^2-\sin^2(\phi))^{m/2}$

Nn::Array{T,1}: integral over $(k^2-\sin^2(\phi))^{m/2}\sin^2(\phi)$

grad::Bool: true for gradient; false for no gradient

s2_grad::Array{T,1}: gradient of $s2$ with respect to $(r,b)$ - this is $S_1 = s_2$ from starry

dsndr::Array{T,1}: derivatives of Green's basis with respect to $r$

dsndb::Array{T,1}: derivatives of Green's basis with respect to $b$

dgdu::Array{T,2}: derivatives $g_n$ with respect to $u_n$

dfdrb::Array{T,1}: derivative of flux with respect to r,b

dfdg::Array{T,1}: derivative of flux with respect to $g_n$

dfdu::Array{T,1}: derivative of flux with respect to u_n

jmax::Int64: maximum number of terms in series expansions of $I_v$ and $J_v$

Mn_coeff::Array{T,3}: coefficients for series expansion of $M_n$

Nn_coeff::Array{T,2}: coefficients for series expansion of $N_n$

ninv::Array{T,1}: inverse of the integers $n$

k2::T: $k^2 = (1-(r-b)^2)/(4br)$

k::T: $k = \sqrt{k^2}$

kc::T: $k_c = \sqrt{1-k^2}$ (unless $k > 1$, then it is $\sqrt{1-1/k^2}$)

kck::T: $k_c k$

kap0::T: $\kappa = \sin^{-1}(k)$ ($=\kappa_0$ in M&A)

pimkap1::T: $\pi-\kappa_1$

sqarea::T: $(1-(b-r)^2)((b+r)^2-1)$

kite_area2::T: $2A_{kite}$

Eofk::T: $E(k^2)$ is complete elliptic integral of first kind

Em1mKdm::T: $(E(m)-(1-m)K(m))/m$ is complete elliptic integral with $m=k^2$

onembmr2:: T: $1-(b-r)^2$

onembpr2:: T: $1-(b+r)^2$

onembmr2inv::T: $1/(1-(b-r)^2)$

onemr2mb2::T: $1-r^2-b^2$

sqonembmr2::T: $\sqrt{1-(b-r)^2}$

fourbr::T: $4br$

sqbr::T: $\sqrt{br}$

sqbrinv::T: $1/\sqrt{br}$

fourbrinv::T: $1/(4br)$

k2inv::T: $1/k^2$

kc2::T: $1-k^2$ (or $1-1/k^2$ if $k > 1$)

sqrt1mr2:: T: $\sqrt{1-r^2}$

den::T: $1/(g_1 + 2/3 g_2)$

third::T: $1/3$

twothird:: T: $2/3$

sqr1mr::T: $\sqrt{r(1-r)}$ if $r < 1$

source